+++ src/runez/__init__.py
@@ -19,7 +19,7 @@ from runez.serialize import read_json, represented_json, save_json, Serializable
 from runez.system import abort, AdaptedProperty, cached_property, PY2, Undefined, UNSET, WINDOWS
 from runez.system import Anchored, CaptureOutput, CurrentFolder, Slotted, TempArgv, TrackedOutput
 from runez.system import decode, flattened, joined, quoted, resolved_path, short, stringified, SYS_INFO
-from runez.system import FallbackChain, first_line, get_version, is_basetype, is_iterable, python_version
+from runez.system import FallbackChain, first_line, get_version, is_basetype, is_iterable
 
 __all__ = [
     "DRYRUN",
@@ -40,7 +40,7 @@ __all__ = [
     "abort", "AdaptedProperty", "cached_property", "PY2", "Undefined", "UNSET", "WINDOWS",
     "Anchored", "CaptureOutput", "CurrentFolder", "Slotted", "TempArgv", "TrackedOutput",
     "decode", "flattened", "joined", "quoted", "resolved_path", "short", "stringified", "SYS_INFO",
-    "FallbackChain", "first_line", "get_version", "is_basetype", "is_iterable", "python_version",
+    "FallbackChain", "first_line", "get_version", "is_basetype", "is_iterable",
 ]
 
 DRYRUN = False